Fix the "delete" button in the table view
In c7fa6c8b5b250c4ca6777af045444c5ffd764863 the code using
`updateFlatStatus` moved to the new file `flatstableline.vue` but the
actual method wasn't moved in the process, yielding an error when one
wanted to use the button.

This patch moves the method to the new file.